Extract Utils To Separate Crate
